requirements.txt
alabaster==0.7.12
asgiref==3.5.2
attrs==22.1.0
Babel==2.10.3
certifi==2022.6.15
charset-normalizer==2.1.0
coreapi==2.3.3
coreschema==0.0.4
coverage==6.4.2
distlib==0.3.5
Django==4.0.6
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
docutils==0.19
drf-spectacular==0.23.1
drf-yasg==1.21.3
filelock==3.7.1
flake8==5.0.4
idna==3.3
imagesize==1.4.1
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.13.0
Markdown==3.4.1
MarkupSafe==2.1.1
mccabe==0.7.0
packaging==21.3
platformdirs==2.5.2
pluggy==1.0.0
psycopg2==2.9.3
py==1.11.0
pycodestyle==2.9.1
pyflakes==2.5.0
Pygments==2.12.0
PyJWT==2.4.0
pyparsing==3.0.9
pyrsistent==0.18.1
pytz==2022.1
PyYAML==6.0
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
snowballstemmer==2.2.0
Sphinx==5.1.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlparse==0.4.2
toml==0.10.2
tox==3.25.1
uritemplate==4.1.1
urllib3==1.26.11
virtualenv==20.16.3
waitress==2.1.2
